fix: more cleanup
