+++ packages/oas/test/analyzer/scope.test.ts
@@ -1,4 +1,4 @@
-import type { OASDocument } from '../../src/types.js';
+import type { OAS31Document, OASDocument } from '../../src/types.js';
 
 import petstore from '@readme/oas-examples/3.0/json/petstore.json' with { type: 'json' };
 import webhooksSpec from '@readme/oas-examples/3.1/json/webhooks.json' with { type: 'json' };
@@ -7,10 +7,8 @@ import { describe, expect, it } from 'vitest';
 import {
   computeOperationScope,
   computeWebhookScope,
-  estimateScopedSize,
   isAncestorOfScope,
   isPointerInScope,
-  toPointer,
 } from '../../src/analyzer/scope.js';
 
 describe('#computeOperationScope()', () => {
@@ -66,20 +64,20 @@ describe('#computeOperationScope()', () => {
 
 describe('#computeWebhookScope()', () => {
   it('should transitively follow `$ref` pointers into referenced schemas', () => {
-    const scope = computeWebhookScope(webhooksSpec as unknown as OASDocument, 'newPet', 'post');
+    const scope = computeWebhookScope(webhooksSpec as OAS31Document, 'newPet', 'post');
 
     expect(scope.rootPointer).toBe('/webhooks/newPet/post');
     expect(scope.reachableRefs).toStrictEqual(new Set(['#/components/schemas/Pet']));
   });
 
   it('should throw if the webhook does not exist', () => {
-    expect(() => computeWebhookScope(webhooksSpec as unknown as OASDocument, 'nope', 'post')).toThrow(
+    expect(() => computeWebhookScope(webhooksSpec as OAS31Document, 'nope', 'post')).toThrow(
       'Webhook `nope` not found.',
     );
   });
 
   it('should throw if the webhook operation does not exist on an otherwise valid webhook', () => {
-    expect(() => computeWebhookScope(webhooksSpec as unknown as OASDocument, 'newPet', 'patch')).toThrow(
+    expect(() => computeWebhookScope(webhooksSpec as OAS31Document, 'newPet', 'patch')).toThrow(
       'Webhook operation `patch newPet` not found.',
     );
   });
@@ -109,36 +107,9 @@ describe('#isPointerInScope()', () => {
 
 describe('#isAncestorOfScope()', () => {
   it('should identify when a scope is nested within a coarser pointer', () => {
-    const scope = computeWebhookScope(webhooksSpec as unknown as OASDocument, 'newPet', 'post');
+    const scope = computeWebhookScope(webhooksSpec as OAS31Document, 'newPet', 'post');
 
     expect(isAncestorOfScope('/webhooks/newPet', scope)).toBe(true);
     expect(isAncestorOfScope('/webhooks/somethingElse', scope)).toBe(false);
-  });
-});
-
-describe('#toPointer()', () => {
-  it('should strip a leading `#` off of a `$ref` pointer', () => {
-    expect(toPointer('#/components/schemas/Pet')).toBe('/components/schemas/Pet');
-  });
-
-  it('should leave an already-plain pointer alone', () => {
-    expect(toPointer('/components/schemas/Pet')).toBe('/components/schemas/Pet');
-  });
-});
-
-describe('#estimateScopedSize()', () => {
-  it('should estimate a nonzero size for an operation that references components', () => {
-    const scope = computeOperationScope(petstore as OASDocument, '/pet', 'post');
-
-    expect(estimateScopedSize(petstore, scope)).toBeGreaterThan(0);
-  });
-
-  it('should not double-count a pointer that appears in both `extraPointers` and `reachableRefs`', () => {
-    const scope = computeOperationScope(petstore as OASDocument, '/store/inventory', 'get');
-    const sizeOnce = estimateScopedSize(petstore, scope);
-
-    // Estimating twice against the same root and scope should be idempotent (this is also what
-    // exercises the memoized size cache).
-    expect(estimateScopedSize(petstore, scope)).toBe(sizeOnce);
   });
 });
